Made ZOrder a read/write property of GraphObject, such that the z position of individual objects may be changed rather than having constant z positions for objects of the same class.
git-svn-id: http://www.bci2000.org/svn/trunk@3356 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/shared/gui/GraphObject.cpp
@@ -36,7 +36,7 @@
 
 using namespace GUI;
 
-GraphObject::GraphObject( GraphDisplay& display, int zOrder )
+GraphObject::GraphObject( GraphDisplay& display, float zOrder )
 : mDisplay( display ),
   mVisible( true ),
   mRectSet( false ),
